add comment and like
add comment and like

+++ mustachajax/src/main/java/com/studymavernspringboot/mustachajax/sbfile/ISbFile.java
@@ -16,8 +16,8 @@ public interface ISbFile {
     String getUniqName();
     void setUniqName(String uniqName);
 
-    Integer getLength();
-    void setLength(Integer length);
+    Long getLength();
+    void setLength(Long length);
 
     String getDescription();
     void setDescription(String description);
