+++ mustachajax/src/main/java/com/studymavernspringboot/mustachajax/sbfile/SbFileDto.java
@@ -1,9 +1,6 @@
 package com.studymavernspringboot.mustachajax.sbfile;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import lombok.experimental.SuperBuilder;
 
 @Getter
@@ -17,7 +14,7 @@ public class SbFileDto implements ISbFile{
     private Integer ord;
     private String fileType;
     private String uniqName;
-    private Integer length;
+    private Long length;
     private String description;
     private String tbl;
     private Long boardId;
